requirements.txt
django 
djangorestframework
django-cors-headers
dj-database-url
